(uboot_2016.07_split) Set correct clock source when resetting BMC
Summary:
Ensure that we are setting the external clock source bit
when resetting the BMC. This has no real functional impact
since (we are resetting the BMC).
But on QEMU which is emulating the writes to this register,
we end up with using PCLK for the boot especially if it involves
resetting during early boot (Common for vboot platforms).

Test Plan:
Boot up tiogapass on qemu.
Flash on real hardware and ensure we did not break anything.

Reviewed By: benwei13

Differential Revision: D17221343

fbshipit-source-id: 4ba3a1be6d