Portwell: Remove workaround for python2+3 coexistence
Summary:
Back in the krogoth days we need to co-exist with some recipes needing
python2 and getting them to support py3. That is changed now.
Everything in openbmc is py3-only and since Rocko, even yocto
recipes have upgraded to using py3.
These changes served a purpose in the krogoth world, but outdated in
the rocko and future world.

Test Plan: Build fbtp. Flash on image and ensure python2 does not exist.

Reviewed By: kawmarco

Differential Revision: D18022199

fbshipit-source-id: bf0ec897d2