Fix reset_usb.sh issue
Summary:
[Galaxy100] Fix reset_usb.sh issue
reset_usb.sh utility, which is shared utility between wedge100 and galaxy100 LC/FC is not working on Galaxy100 LC/FC
The reason was because the sysfs name for this feature in CPLD driver is different in Galaxy100 LC/FC.
It should be usb_hub_rst_n as in Wedge100's CPLD driver, but it was usb_hub_rst (missing _n)

This diff fixes this issue. The fix has been tested in a unit that had an issue in the fleet.

Test Plan: P120705887

Differential Revision: D18263398

fbshipit-source-id: cbb8a3d4c5