openbmc-init-build-env: zsh fixes
Summary:
The machine/meta search code was relying on bash word splitting behavior
and the `==` operator.  Set the zsh option `sh_word_split` to enable
bash-like splitting, when zsh is detected.  Reduce the `==` to an `=`
because this is the correct string equality test in a single `[` test.

Test Plan:
Source in both bash and zsh where $1 is not a fully qualified meta, such
as `source openbmc-init-build-env angelslanding build-fbal`

Reviewed By: amithash

Differential Revision: D18694498

fbshipit-source-id: a890b989b9