pal: Add a common interface to send a power-button-override
Summary:
Add a platform method to initiate a power-button override which is common across
server platforms with the CPU hosted on the motherboard. This method is expected
to be initiated after restoring the BIOS SPI flash ownership back to the PCH/CPU after
a BIOS flash upgrade.

This method has different names across tiogapass, sonorapass and angelslanding. This
change just unifies it and puts the API in common obmc-pal.

Test Plan:
Build tiogapass, sonorapass, angelslanding.
(Functional test bypassed due to no functional change)

Reviewed By: benwei13

Differential Revision: D18711365

fbshipit-source-id: cdd915ba2f