spi-aspeed: add transfer-one function
Summary:
The spi subsystem prefers devices give a 'transfer-one' function to
perform a single SPI data operation.  There is a higher-level function,
'transfer-one-msg', which is implemented using pieces such as
transfer-one and chip-select and allows fine-grained control of the
chip-select between pieces of a transfer.  This is necessary for some
SPI slave device drivers, such as TPMs.

Test Plan: TPM testing on sonorapass

Reviewed By: tomrepo

Differential Revision: D18966748

fbshipit-source-id: aefd692356