tpm: spi: support half-duplex masters
Summary:
The TCG SPI spec does not officially support half-duplex masters
because the TPM chips can send a "busy wait" status in the 4th
byte (and subsequent bytes until the busy state is completed).

The address used to read the DID/VID happens to have a 0 in the
4th byte and is safe to read at any time.  On half-duplex masters,
we can insert a DID/VID command to read and handle the "busy wait"
state from a TPM before sending the desired command.

This has been tested on the AST2500 (master) and Infineon 9670 (slave).

Test Plan: TPM testing on sonorapass

Reviewed By: tomrepo

Differential Revision: D18966746

fbshipit-source-id: 81ef21245b