sonorapass: disable TPM and enable BIOS flash
Summary:
Change the device tree for Sonora Pass to use the aspeed-smc MTD driver
instead of the generic ast2500-spi-master driver.  This re-enables the
BIOS MTD device for OOB BIOS update and disables the TPM.

There are multiple aspects of the BIOS update path that do not currently
work with the ast2500-spi-master path and will need additional
development work.  We need the OOB BIOS update to work for EVT builds.

Test Plan: Boot on SonoraPass, manually set mux GPIO and re-probe MTD.

Reviewed By: amithash

Differential Revision: D19304886

fbshipit-source-id: dffd77f609