travis-ci: reduce git-clone depth
Summary:
On some repositories we include the kernel tree, so cloning takes
a long time (over 1 minute) and is duplicated for every job.
Limit the clone depth in effort to speed that up.  This saves us
approximately 15 minutes total of Travis compute time.

50 commits should be plenty sufficient to find the common ancestor
or else this commit under test really needs to be rebased.

Test Plan: PR to openbmc/openbmc to trigger Travis

Reviewed By: benwei13

Differential Revision: D19431713

fbshipit-source-id: 2166170c84