gpv2: add machine-image symlink
Summary:
In order to run TravisCI on gpv2, we need a match between the machine
name and the image name.  We effectively run:
```
    source openbmc-init-build-env ${BOARD}
    bitbake ${BOARD}-image
```

gpv2 is found by openbmc-init-build-env as fbgp2.  Add a symlink for
fbgp2-image to fby2-gpv2-image so both parameters can match.

Test Plan: Manually run the above commands.

Reviewed By: sammyblueeyes, benwei13

Differential Revision: D19559365

fbshipit-source-id: 9d519bccbb