common: tzdata: reduce installed timezones to 'core'
Summary:
The tzdata recipe is divided up into a number of packages that
installs MBs worth of timezone data from all over the world.
We only use a small number of these timezones, which are included
in the "core" set.  Modify the RDEPENDS so that only the core set
is installed.

Test Plan: Built and booted on Sonora Pass

Reviewed By: tfangit

Differential Revision: D19265612

fbshipit-source-id: 6fb7071c04