fw-util: partial lexicographical ordering of FRUs and components
Summary:
When a component/FRU ends with a numeric value, compare
lexicographically as long as the strings do not end with
a numeric component. If they do, compare them lexicographically
till the numeric component which it itself compared numerically.

In the new scheme
cevice < device
device4 < device10

Test Plan:
Test on GPv2 which has device0-device12 where the difference
is easy to test.

Reviewed By: benwei13

Differential Revision: D20205196

fbshipit-source-id: f8713791a6