fw-util: Warn on deprecated utility usage
Summary:
Passing the component name as if it were an option to the utility was
added to be backwards compatible to the old way. Now that we have both
available for a year now, it is safe to start marking the old as deprecated.

We should eventually remove the support. But this gives any users heads up.

Test Plan:
root@bmc-oob:~# fw-util bmc --version --bmc
WARNING: Passing component as --bmc is deprecated
         Use: --version|--update bmc instead
BMC Version: fbtp-9b4c123717-dirty

Reviewed By: benwei13

Differential Revision: D20566224

fbshipit-source-id: c6f6f13202