sonorapass: fan-util: set fan zones to 2
Summary:
This commit is effectively a revert of D20470590.

The fan control set was changed to have 16 fans, but the fans cannot
be changed independently.  This leads to a confusing user interface
where changing one fan changes others.  Revert `fan-util --set` to
operate on zones rather than individual fans.

Test Plan:
Run `fan-util --set 30 0` and `fan-util --set 40 1` and
confirm expected results (half fans are set to 30% and half to 40%).

Reviewed By: vineelasmile

Differential Revision: D20706435

fbshipit-source-id: 986b92fe3b