fan-util: improve usage display
Summary:
The current usage of `fan-util` on Sonorapass is:
```
Usage: fan-util --set <[0..100] %> < Fan# [0, 1] >
       fan-util --get < Fan# [0..15] >
```

Sonorapass has 4 physical fans, each with 2 rotors, where each
fan is assigned in pairs to a fan-controller.  Thus there are
8 tach sensors assigned to 2 zones.  (Physically, the system
could have 4 additional fans added, but these are not currently
present.)

Adjust the usage of `fan-util` so that when the number of PWMs
do not match the number of Tachs, we use the Zone/Tach
terminology instead of the inconsistent Fan/Fan.

The new usage will appear as:
```
Usage: fan-util --set <[0..100] %> < Zone# [0, 1] >
       fan-util --get < Tach# [0..15] >
```

Test Plan: Build and run on sonorapass

Reviewed By: vineelasmile

Differential Revision: D20706433

fbshipit-source-id: 10e45a7cbc