openbmc-gpio: fix python3 DEPENDS to be python3-setuptools
Summary:
Numerous machines failing with a bitbake error:
ERROR: Nothing PROVIDES 'python3-distribute' (but
common/recipes-utils/openbmc-gpio/openbmc-gpio_0.2.bb DEPENDS on or
otherwise requires it).

This build regression was introduce by D20807667, which fixed the
DEPENDS line in the recipe but the line already contained an invalid
package dependency.

This package is using what is contained in python3-setuptools in
order to install itself (setup.py file), so update the dependency
to match.

Test Plan: Passes CIT build.

Reviewed By: amithash

Differential Revision: D20838372

fbshipit-source-id: e0ed9a0b2b