common: handle versioned libraries across python
Summary:
The python ctypes.CDLL call requires a specific shared library
file name be passed as a parameter.  As we move away from
non-version shared libraries, as is Yocto convention, we need
to ensure the parameter to CDLL includes the version.  Pre-emptively
add this now so we do not run into problems in the future as we
refactor to versioned shared libraries.

Also fix a large number of python lint errors.

Test Plan:
Test on Sonorapass loading all libraries in python by versioned name.
```
>>> CDLL("libpal.so.0")
<CDLL 'libpal.so.0', handle a50e30 at 0x76968630>
>>> CDLL("libgpio-ctrl.so.0")
<CDLL 'libgpio-ctrl.so.0', handle a66800 at 0x769685f0>
```

Reviewed By: sammyblueeyes

Differential Revision: D20745610

fbshipit-source-id: be6be5240c