fan-ctrl: clean up bitbake warning for missing file
Summary:
Observing the following warning when building some machines:
    common/recipes-core/fan-ctrl/fan-ctrl_0.1.bb: Unable to get checksum
    for fan-ctrl SRC_URI entry fand.cpp: file could not be found

D20878918 moved the file from the common layer into the machine-meta
because the file was only applicable to a small number of machines, but
the recipe was unchanged.  When bitbake parses the recipe files for
machines without this file it gives the above QA warning.  Move the
SRC_URI directive from the .bb to the .bbappends.

Test Plan: Build tiogapass, confirm bb warning is gone.

Reviewed By: tao-ren

Differential Revision: D21183881

fbshipit-source-id: fd837b8bd1