jf: Fix issue where build was not triggered for platforms
Summary:
If a diff changed files affecting platforms X, Y and Z,
then some files affect some regex is the correct policy
for the filtering to trigger a build job for X Y and Z.

This issue was discovered when I noticed a large diff
changing files in each platform: D21209027 was building only fby2
based platforms.

Test Plan:
Bug visible on V1 of D21207976 and D21209027
Added this diff to V2 of D21207976 to ensure that build of
all platforms was initiated.

Reviewed By: williamspatrick

Differential Revision: D21208971

fbshipit-source-id: 33e2129639