Correct System Assemply Part Number Parsing in we/se/pe utils
Summary:
Facebook MQE team uncovered a bug in weutil/seutil/peutil, where we were incorrectly parsing SYSTEM ASSEMBLY PART NUMBER field in FBOSS platforms.
It should be XX-XXXXXXX-XX or XX-XXXXXX, but we were parsing it as XXX-XXXXXX-XX or XXX-XXXXX. This is not correct (the place we insert the first hyphen is wrong)
This diff corrects that issue.

Test Plan:
This value is not used by FB Infra, nor is in SeRF. It's used only in some vendor teams. Thus verifying the function of this diff should be good enough.

Before fix :
P132683865

After fix:
P132684197

Reviewed By: williamspatrick

Differential Revision: D21911481

fbshipit-source-id: 42ac84879b