fby3: libpal: update parameter for adm1278 reading (#1101)
Summary:
-The Rsense of the 2nd source of adm1278 is changed in DVT stage.
A power team measured the accuracy of sensors of adm1278 and found that sensors are inaccurate.
In order to make sensors are more reliable and not exceed a percentage error 2%.
It provides a new parameter for those sensors. The 2nd source is verified and passed now.
Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/1101

Test Plan: -built code(pass)

Reviewed By: zhdaniel12

Differential Revision: D22202767

Pulled By: benwei13

fbshipit-source-id: 52aecc0b80