u-boot v2019.01: fbep: Disable to reset the PWM controller (#1781)
Summary:
The OAMs generate heat if server is on-line.
However, BMC will be held if PFR validation is on-going, the default PWM is not enough to cool the OAMs
This patch disabled the reset of PWM contoller so that fans will keep the last duty cycle we set.

Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/1781

Test Plan: Build fbep pass

Reviewed By: benwei13

Differential Revision: D22458204

fbshipit-source-id: d5a6554277