vboot, u-boot: Fix issue when new u-boot has to load old kernel-fit
Summary:
Early implementation of vboot would pass in dual_flash == 1 for
verified boot based systems and dual_flash == 0 for non-vboot.
Kernel would explicitly check for dual_flash == 1 to indicate vboot support.
Lets call it IMAGE_A.

In order to support platforms with dual flash but no verified boot,
this was enhanced to pass dual_flash == 0 for single flash,
dual_flash == 1 for dual flash and dual_flash == 2 for dual flash
with verified boot. Lets call this IMAGE_B.
The change was done to be backwards compatible and IMAGE_B::kernel
would work being loaded by IMAGE_A::uboot.

Unfortunately, when IMAGE_B::uboot loads IMAGE_A::kernel then
dual_flash==2 is passed to the kernel which ends up with a bug
since it is checking for dual_flash == 1 for verified boot or not.
Under normal flow, this can never happen since the RW image is always
newer than the read-only (golden) recovery image. But if the kernel
fit is corrupted, then u-boot on the RW flash1 is the one which loads the kernel-fit
from flash0 -- Thus causing the issue.

This is part 1 of the change to revert u-boot to pass dual_flash==1
to the kernel irrespective of whether it is vboot or not.
This should make u-boot compatible with the older kernel (IMAGE_A).

Thankfully this fix is in the writable u-boot section.

Test Plan:
1. Basic flash on tiogapass and ensure vboot succeeds.
2. Corrupt fit (By erasing a block in the fit area.) and reboot --
   we should boot into recovery and it should still be able to
   show both the flashes in /proc/mtd.

Reviewed By: benwei13

Differential Revision: D22479794

fbshipit-source-id: 431f86a92f