Adapt fan tests to systemd
Summary:
Fan-related tests on Wedge100 rely on runit-based daemons
that we are converting to systemd units. Detect systemd and when
available use the systemd equivalents for managing these daemons.
it

Test Plan:
Ran on a wedge100 without systemd, all tests pass. Booted a
wedge100 with systemd and tests still pass.

Reviewed By: sammyblueeyes

Differential Revision: D22636093

fbshipit-source-id: 7b9db5c360