pfr: Do not expose the first 64k of the MTD partition
Summary:
PFR detects access to the early parts of MTD as a reset.
Do not allow access to this region through the standard MTD partition so user space applications can continue to function without needing to be aware of the HW quirks.

Test Plan:
Build, sign and flash on PFR enabled Y3.
Ensure that everything works, and the BMC does not reset when accessing the bmc node of rest-api which continues to have the u-boot version correctly displayed.

Reviewed By: zhdaniel12

Differential Revision: D22781415

fbshipit-source-id: c467c612a0