hostname: Allow platforms to set the primary network interface
Summary:
eth0 while a good default is not always the primary network
interface for every platform. This change allows platforms
to override the default interface over which we need to
query the hostname.

Test Plan:
Flash on yosemite V2, delete /mnt/data/hostname and reboot.
After boot check if the hostname is correctly updated and `ps` does not show hostname.sh (It terminates, does not linger forever).

Reviewed By: sammyblueeyes

Differential Revision: D22914436

fbshipit-source-id: 6fb05d5f14