angelslanding: Set the primary network interface as br0
Summary:
On Angelslanding, the primary interface to public
is br0 and not eth0. This change sets hostname
scripts to query for current hostname on that interface.

Test Plan:
Flash on angelslanding, delete /mnt/data/hostname and reboot.
After boot check if the hostname is correctly updated and `ps` does not show hostname.sh (It terminates, does not linger forever).

Reviewed By: zhdaniel12

Differential Revision: D22914437

fbshipit-source-id: 95ff2c57ba