Retry getting hostname if the first attempt fails
Summary:
Retry up to 10 times with a 15s interval to fetch
the hostname of the BMC. This ensures enough time
in BMCs where the network setup is complex and can
take longer time than the initial 15s.

Test Plan:
1. Regression testing on fby2
remove /mnt/data/hostname and flash on fby2. After boot
ensure that hostname is populated and `ps` does not show
hostname.sh running.

2. Repeat Step 1 on angelslanding
3. Repeat step 1 on emeraldpools

Reviewed By: benwei13

Differential Revision: D22958798

fbshipit-source-id: c7b8b85867