flashy: Add compatible version mapping for fby3vboot2 & fby3pvt
Summary:
fby3vboot2 and fby3pvt are temporary names used to differ these builds
with current fby3 official release.

When the YV3 program make the decision to go with either vboot or pfr,
we will switch the corresponding build to offical release and remove
the other one.

If it decide to go with PFR, the fby3pvt will become the offical release
of fby3, and fby3vboot2 will be removed. If it is decided to not go
with PFR, fby3vboot2 will become official release of fby3, and fby3pvt
will be removed.

Test Plan: Unit tests (`go test ./...`)

Reviewed By: kawmarco

Differential Revision: D23051304

fbshipit-source-id: 7fe0650f71