jtag: aspeed: add AST2600 support (#45)
Summary:
Enhance the JTAG driver to support the AST2600.

On AST2600, the AST2600-specific JTAG pins are enabled
using an SCU enable pin register.

Also for AST2600, the meaning of the SCU RESET_JTAG
register is different. To set a bit, the SCU_RESET_JTAG
register is used. To clear a bit, an additional clear
register is used.

Test Plan:
These changes were testing with additional changes on
the yamp and elbert platforms to verify that FPGAs could
be updated using the jam utility.

These specific driver changes were tested in isolation
by building an image for a yamp platform with kernel 5.6
and verifying that the JTAG device was present.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/45

Reviewed By: joancaneus

Differential Revision: D23093876

fbshipit-source-id: e9f91a52fb4f92b6741fa5ada5c30a8509522baa