ARM: dts: aspeed-g6: enable jtag controllers
Summary: Enable JTAG controllers in aspeed-g6.dtsi.

Test Plan:
These changes were testing with additional changes on
the yamp and elbert platforms to verify that FPGAs could
be updated using the jam utility.

These specific driver changes were tested in isolation
by building an image for a yamp platform with kernel 5.6
and verifying that the JTAG device was present.

Pull Request resolved:
https://github.com/facebookexternal/openbmc.arista/pull/45

Reviewed By: joancaneus

Differential Revision: D23093876

fbshipit-source-id: b308f77f85