Handle gracefully reload
Summary:
In this service we trigger a reload by killing it with
`-HUP`. In some platforms this may kill the service, in that case we
let systemd fall to a normal restart.

Test Plan:
Booted an image and ran `systemctl reload restapi`. It
kills the process with HUP as expected:

```
[[0;1;[[0m restapi.service - OpenBMC REST API
   Loaded: loaded (/lib/systemd/system/restapi.service; enabled; vendor preset: enabled)
      Active: [[0;1;32mactive (running)[[0m since Wed 2020-09-09 09:30:39 PDT; 8min ago
        Process: 3219 ExecReload=/bin/kill -HUP $MAINPID (code=exited, status=0/SUCCESS)
         Main PID: 1031 (python3)
            CGroup: /system.service.slice/restapi
                       ]]]]]]
      32m]]
```

Reviewed By: deathowl

Differential Revision: D23600657

fbshipit-source-id: 336b2629d7