Adapt recipe to systemd
Summary:
Add a unit for this service, and install either the systemd
or the SysV way, depending on whether the platform has it enabled.

Test Plan:
Confirmed on angelslanding the runit service is gone, and
that the systemd unit is there.

Disabled systemd and re-built. The runit services reappear.

Note: At this stage I am only confirming that all files end in the
right places. Testing that the daemons work on Systemd will happen
later.

Reviewed By: doranand

Differential Revision: D23789072

fbshipit-source-id: 705aaf697c