openbmc-init-build-env: short-circuit distro search on match
Summary:
The distro override loop has some logic to allow matches on
machine prefixes.  This allows different configurations of the
same machine to have a single distro override setting, such as 'fby3'
and 'fby3s'.  The problem comes from systems which are themselves
named by a prefix: { wedge, wedge100, wedge400 }.

The older machine is likely to have an older distro and be later in
the array.  The simplest solution is to short-circuit on first match
so that 'wedge400' matches and exits.  Without this, 'wedge400' attempts
to build with 'rocko' instead of 'warrior' by default.

Test Plan: Source for wedge400 and confirm warrior is chosen.

Reviewed By: tao-ren

Differential Revision: D24489122

fbshipit-source-id: b06c3d86f3