circleci: reduce cases of full machine rebuilds
Summary:
I've noticed the CircleCI jobs are very infrequently detecting that
the ODM pull requests are indeed PRs and reducing the target machines
to just the affected set.  The is because the typical ODM development
model is that they push to a named-branch and then create the PR.  The
CircleCI sees a "branch update" and kicks off a full branch build before
the developer is able to create the PR.

Try to identify named-branches that will eventually become PRs by
treating any CIRCLE_BRANCH not in { 'master', 'helium' } as a PR.

Test Plan:
Ran check-build-required manually with some environment variables
from CircleCI logs to confirm desired behavior.

Reviewed By: benwei13

Differential Revision: D24567603

fbshipit-source-id: 5929229fd6