Support user specified max log size on OpenBMC mTerm
Summary:
This change adds an optional 4th perameter to the mTerm_server that
allows a user to specify the max log size. If the max log size is bellow
the default value or above the max value it will be set to the default.

The current default min is 300K
The current max is 10 MB

Test Plan:
{F343554167}
mTerm_server running with no arguments will stop the logging at 300k as per the original version.
 ---
{F343554262}
Log is now longer than the default minimum.

{F343556479}
Running with log set to 10MB

 ---
Example comparing output of `sol-util slot2 --history` with terminal of slot2

{F343776123}

 ---
New output of mTerm_server util.

{F343776284}

Reviewed By: benwei13

Differential Revision: D24769794

fbshipit-source-id: 9de459a7f2