Disable i2c devices when not in BMC mode (#106)
Summary:
When the system is not in BMC mode, there's a potential for i2c devices to
conflict with accesses from the CPU. To be safe, don't set up these devices in
the BMC when the system is not in BMC mode.

In addition, disable the logic that brings the TH4 in and out of reset.

This is mainly to prevent issues if the BMC is powered on when the CPU
is running EOS.

Testing Performed:
1. Booted the BMC and verified that there are appropriate messages in
   the boot log indicating that device setup is skipped when not in BMC mode.

2. Verify when in BMC mode the BMC boots up normally and i2c devices
   are setup.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/106

Reviewed By: tao-ren

Differential Revision: D24869569

fbshipit-source-id: cd58002089