rest-api: remove native dependency on libgpio-ctrl
Summary:
None of the rest-api testcases rely on the python module supplied by
libgpio-ctrl.  It is desirable to remove libgpio-ctrl from the native
context because it doesn't successfully build everywhere.  Remove the
DEPEND and RDEPEND and add it specifically to the RDEPEND_class-target
so it is only depended on in target contexts.

Test Plan: `bitbake` successful, which runs tests on `rest-api`.

Reviewed By: benwei13, deathowl

Differential Revision: D25133856

fbshipit-source-id: ca21e5d0de