libgpio-ctrl: remove native context
Summary:
This is not needed, so remove it.  It also fails to build in some cases
where the host compile (`gcc`) is an older version that does not default
to at least C++11.  In that case, `gtest` is compiled with C++03 support
and `libgpio-ctrl` is compiled with C++17 support, and the two do not
link together successfully.

Test Plan: `bitbake` build an image.

Reviewed By: benwei13

Differential Revision: D25133855

fbshipit-source-id: 0d51b72f0b