Cache the result of get_internal_switch_config
Summary:
Add a local decorator for caching the results and apply it to
internal switch config *for wedge-100 platform ONLY* for now. If we
decide this is a viable approach - we can generalize this code and apply
to other platforms.

Test Plan:
Flashed the fw on a test device - caching works as shown by
time it takes to run the first and then subsequent commands.

Speed up is ~5x

  nikoladj@devvm1524:wedge100  (helium)$ time curl fboss8382003-oob.snc1:8080/api/sys/firmware_info/internal_switch_config
  {"INTERNAL_SWITCH_CONFIG": "6a5180282641e9683099a9030f26f954ab98ee3b1ebdcd774ad50b58a4f6a7ef"}
  real    0m13.488s
  user    0m0.007s
  sys     0m0.005s
  nikoladj@devvm1524:wedge100  (helium)$ time curl fboss8382003-oob.snc1:8080/api/sys/firmware_info/internal_switch_config
  {"INTERNAL_SWITCH_CONFIG": "6a5180282641e9683099a9030f26f954ab98ee3b1ebdcd774ad50b58a4f6a7ef"}
  real    0m0.348s
  user    0m0.006s
  sys     0m0.007s
  nikoladj@devvm1524:wedge100  (helium)$ time curl fboss8382003-oob.snc1:8080/api/sys/firmware_info/internal_switch_config
  {"INTERNAL_SWITCH_CONFIG": "6a5180282641e9683099a9030f26f954ab98ee3b1ebdcd774ad50b58a4f6a7ef"}
  real    0m0.343s
  user    0m0.005s
  sys     0m0.007s

Reviewed By: vineelasmile, deathowl

Differential Revision: D25231843

fbshipit-source-id: b4092ccb55