Add unittests for kv_cache_asyn
Summary:
The caching decorator tested here was added in a related
commit, so this commit addes unittest for it (it was previously tested
in a built image). They are currently confined to wedge-100 as this is
the only place where we use the caching functionality for now - it might
be moved out to a more common place in the future. The test was written
in a way that is agnostic to what the actual decorated function does.

It also adds a "fake" kv module as we can't import it during unittests

Test Plan: bitbake rest-api-native  -c do_unit_test

Differential Revision: D25279977

fbshipit-source-id: 5ab1b11f67