Fix rest endpoint /api/sys/bmc crash on Wedge400 and C
Summary:
[FBOSS] Fix rest endpoint /api/sys/bmc crash on Wedge400 and C
Wedge400 and 400C ANR fails because rest API handler for /api/sys/bmc crashes.
This was because there was not enough error handling in the logic to read u-boot
version. Thus the handler could not recover, when it somehow failed to read
image meta-data information directly, in order to get u-boot version.
This diff properly handles all the issue in reading u-boot version, and
gracefully set uboot version as "NA", in case uboot version could not be read.

Test Plan:
Before : P153805968
After : P153806026

Reviewed By: williamspatrick

Differential Revision: D25432545

fbshipit-source-id: ca7b8fe255