Introduced common_auth.Identity, capture host identity
Summary:
Some endpoints (such as `POST /api/sys/fscd_sensor_data`) should only be authorised if the request comes from certain hosts (such as the managed server).

In this case, we need a reliable way to identify the request's client. This diff captures the client identity as either the host string as it appears in the TLS client cert or the source ip address.

Test Plan: Unit tests, CIT, checked REST api on a test wedge400c. More thorough testing on incoming changes.

Reviewed By: deathowl

Differential Revision: D25495350

fbshipit-source-id: 4b912e4a32