fbep: Lock ASIC bus without non-blocking while reading vBIOS version (#2266)
Summary:
There is a chance that the bus was held by sensord over 300 ms since 2 OAM shared 1 SMBus.
If we don't want to see "Not available" showed by fw-util while reading vBIOS version due to timeout
we have to try to lock the bus without non-blocking instead.

Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/2266

Test Plan:
Build pass and test on fbep (with 2 Nvidia OAMs)

Reading vBIOS with 1000 times without error or "Not available"

for i in {0..1000};
do
  ret=$(fw-util asic --version | egrep "(Not|NA|Error)")
  if [[ "$ret" != "" ]]; then
    exit -1
  fi
done

Reviewed By: benwei13

Differential Revision: D25958867

fbshipit-source-id: a7ab1c1c90