Adding Autonomous Crash Dump Source to ODM Repos.
Summary:
This commit adds the proprietary source to both Quanta and Wiwynn repos.

This initial commit contains the code ported to work with our platforms
by the ODM. Future commits will work on decoupling our platform specific
code from Intel's code in order to make merging changes easier.

Code in any of the proprietary directories should NOT be moved to
another directory for any reason. If the code is moved to another
directory and committed it will then be shipped publicly.

Test Plan:
1. Make sure Shipit only tries shipping to the correct repos.
2. Builds on all affected platforms (fbal, fby3)

Reviewed By: benwei13

Differential Revision: D26133838

fbshipit-source-id: ccc9b74c68