sensor-util: Hide absent FRU errors when displaying all
Summary:
While it should be an error when user wants to view the sensors
of an absent FRU, it should not be an error when the user
wants to display sensors from all FRUs and one or more FRUs
are absent. In a marginally complicated system this is not
a failure (All cards need not be present in all configurations!).

This changes the code to make sure we do not print an error
and fail the call if the user is printing all sensors if
the FRU happens to be absent.

Test Plan:
Tested on tiogapass with riser_slot2 absent.
1. Ensure that asking for an absent FRU prints the correct error message
and fails the command.
```
root@ac-fb-08:~# sensor-util riser_slot2
riser_slot2 is not present!

root@ac-fb-08:~# echo $?
255
```
2. Ensure that running sensor-util all succeeds.
```
root@ac-fb-08:~# sensor-util all > /dev/null
root@ac-fb-08:~# echo $?
0
root@ac-fb-08:~# sensor-util all | grep "not present"
root@ac-fb-08:~#
```

Reviewed By: GoldenBug

Differential Revision: D26343973

fbshipit-source-id: cf6b3bbced