pal: Remove duplicated declarations in platform pal.h
Summary:
This is more of a code cleanup than a functional change.

There are duplication of declarations of PAL methods in
platform layer's pal.h. These are already declared in obmc-pal.h
and should not be duplicated.

This would potentially make changes to prototypes easier
since along with changing every pal.c we would not need to change every
pal.h along with it.

Test Plan:
Build all the platforms!
Since no functional bits are touched, we do not need to do
any runtime tests. If all platforms build, we are good.

Reviewed By: tao-ren

Differential Revision: D26377949

fbshipit-source-id: 925cf09745