Fix Small Buffer Overrun in nvme_part_num_decode()
Summary:
While testing possible static analysis tools one of the tools revealed
that nvme_part_num_decode() was setting a null terminator for a string
buffer +1 past the length of the buffer.

This was fixed by using the defined buffer length and making the buffer
"length + 1". A new definition for the key length was also created and
the assignment of that key was changed from sprintf to snprintf.

(I don't see a reason to not almost universally use snprintf over
sprintf)

Test Plan: Builds on systems.

Reviewed By: benwei13

Differential Revision: D26474351

fbshipit-source-id: 52f250ec07