mask bind service too
Summary:
A service named "bind" is autogenerated for /etc/init.d/bind, and systemd tries to start it at boot time.

We'll turn off sysvinit compat at some point in the future, and this init script won't be a problem any more.  But until then we should mask the service so we don't try and start it at boot.

Test Plan:
Build a new image and boot it on test BMC.  Check that bind service is masked:

```
root@bmc:~# systemctl status bind.service | cat
● bind.service
   Loaded: masked (/dev/null; bad)
   Active: inactive (dead)
```

Reviewed By: benwei13

Differential Revision: D26666808

fbshipit-source-id: 87e3593c7a