cm-util: Fix wrong display of link-local address
Summary:
Looks like std::setfill() reverts to default as soon as it
is used for the first number/output.
So
```
std::cout << std::setfill('0') << std::setw(2) << std::right << std::hex
<< 0xfe << int(6) << std::endl;
```
ends up printing fe6 instead of the correct fe06 because of the iomanip
is not applicable to the next.
This change just splits up and have each line print just one digit.

Test Plan:
Build cm-util, copy the ipk to the problem system and install and re-run
cm-util. The output should be correct

Reviewed By: GoldenBug

Differential Revision: D26896490

fbshipit-source-id: 4d5cd72262