emeraldpools: Remove dummy source
Summary: At one point in time, nic-components were enabled (constructed and added to the firmware list) by default. So, platforms like emeraldpools which do not have a NIC had to workaround this to disable the component. We no longer have NICs enabled by default, so this override is unnecessary.

Test Plan: Build emeraldpools

Reviewed By: benwei13

Differential Revision: D26949333

fbshipit-source-id: 5d7db380fb