rackmond: Add support for monitoring SOH timestmap
Summary:
The PSU monitors Modbus 0x121 or SMBUS 0x32 low byte for the count of
successful SOH test. When the number increases, PSU will record
timestamp(0x12A) to 0x131. If wall clock time stays 0 when updates, then
0x131 will retain its current value.

This is safe to read for PSUs which might not support it.

Test Plan: #doitlive

Reviewed By: tao-ren

Differential Revision: D27479555

fbshipit-source-id: f3fb636e31