Increase the shell command timeout of PIMSERIAL endpoint to 30 seconds
Summary:
[Minipack1] Increase the shell command timeout of PIMSERIAL endpoint to 30 seconds
With current timeout value (default value, which is 20), the very first execution of
dump_pimserial.sh will always timeout, as it takes more than 20 seconds to read all of them
for the first time. (after that, we use cached value, so we don't see any more issue)

This was causing our provisioning of Minipack1s to fail. So increasing the timeout value.

Test Plan: Before and After : P368508438

Reviewed By: tao-ren

Differential Revision: D27514679

fbshipit-source-id: 4774737352