netcastle: make ptest-runner build common
Summary:
ptest-runner can be used to call bitbake with the `--force-build`
option, but when we do that we end up with different logging than
the rest of the machines.  Refactor the bmc_build code that uses
ptest-runner so that machine build is separate from ptest execution
and machine build is common with the non-ptest code path.

Test Plan: Confirm ptest passes in CIT.

Reviewed By: amithash, benwei13

Differential Revision: D28394751

fbshipit-source-id: 5dd06da351