ELBERT: Verify images read by fpga_util.sh (#198)
Summary:
Although fpga_util.sh program verifies images written and retries on
failures, fpga_util.sh read does not verify the contents read.
As a result, it is possible that an image read may be corrupt and
if this file is later used to write into the flash, then the flash
will be corrupt as well.

The solution is to verify the image after reading and report an error
if the image read does not match the image on flash. That way, any
scripts using fpga_util.sh read will not inadvertantly make use of
the image.

Testing performed:

Ran a script for 5000 iterations that repeatedly read from the SPI
flash. The test script then compared the bytes to a known good image.
The test script detected 122 cases where fpga_util.sh read reported
that the image was corrupt. However, in cases where fpgautil.sh read
reported success, the test script never detected a corruption.

In addition, verified that the following scripts were still working:
- spi_pim_ver.sh
- th4_qspi_ver.sh

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/198

Reviewed By: mikechoifb

Differential Revision: D29944971

fbshipit-source-id: aec02d0aa3