added LCR to sensor events
Summary: This adds another subtype to sensor events that asks for the bound type, either UCR or LCR. If none is given, then it defaults to UCR because that is the most common event type. It no longer needs a payload field, but will not throw an error if one is given. Additionally, because there is the defaulting to UCR, any existing sensor events will still be interpreted correctly.

Test Plan:
I tested with events 0, 1, and 10 from the conf file in this diff:
D29970084

Please see this paste with verbose output from testing:
P433351047

Reviewed By: benwei13, lhl2617

Differential Revision: D29970047

fbshipit-source-id: d6a90ab8b0