distro: remove image-mklibs support
Summary:
The image-mklibs bbclass has been removed from later versions of
Yocto and the `mklibs` tool won't even compile with recent host
GCC.  We have always had this bbclass added, but never enabled
because the MKLIBS_OPTIMIZED_IMAGES variable was never set.  Since
we're already not using it on Rocko, it is safe to remove it
everywhere.

Test Plan: Build northdome.

Reviewed By: GoldenBug

Differential Revision: D30083296

fbshipit-source-id: da86485c40