northdome: speed up power-on wait
Summary:
When the slot power-on fails, the power-on script waits for a timeout
rather than continuing on in the face of error.  On QEMU, the power
on always fails, so the boot up hangs for a very very long time.

Switch the retry loop to a simple `wait` call to wait for all power-util
child processes to have completed.

Test Plan: QEMU boot is faster on Northdome.

Reviewed By: GoldenBug

Differential Revision: D30083294

fbshipit-source-id: efc653320a