Remove exception if libpal fru is not present
Summary: I noticed on a wedge400 that fru_name_map returns an fru_name thats not found on the second check via pal.pal_is_fru_prsnt. Previously, we were raising an exception in such a case which isn't a good idea. We should probably return empty response for an fru thats not present so that we can still continue with valid frus in the fru_name_map.

Test Plan:
Curling Redfish Thermal endpoint on a wedge400
P437014606

Reviewed By: deathowl

Differential Revision: D30218543

fbshipit-source-id: eca797c692