rocko, zeus: workaround for absent version-less python
Summary:
CentOS8 doesn't have a version-less python executable.  In our
devservers we workaround that by adding an appropriate 'alternative'
so that python maps to python2.  On Sandcastle machines we don't have
this luxury and need an alternative.  When our jobs get assigned to
a new CentOS8 instance, we get the following build failure:

```
OpenEmbedded requires 'python' to be python v2 (>= 2.7.3), not python v3.
```

Install a simple 'python' shell script that acts as an indirection to
the python2 found in the $PATH.  This is similar to what an alternative
in the system would do.

Add a small update to $PATH when setting up the Yocto environment if
`$distro/workaround-bin` exists, so that this version-less python is
found by bitbake (and the oe setup scripts).

Test Plan:
Executed the setup script in a Docker environment with the version-less python workarounds.
```
pokyuser@84718712eb72:/workdir/local/builds/build-fbtp$ which python
/workdir/local/sync/fbopenbmc/yocto/rocko/workaround-bin/python
pokyuser@84718712eb72:/workdir/local/builds/build-fbtp$ python --version
Python 2.7.17
```

Reviewed By: zhdaniel12, GoldenBug

Differential Revision: D30462511

fbshipit-source-id: fa7b15517d