openbmc_version: fix potential PSEUDO violation
Summary:
We are regularly seeing failures in developer builds such as:
```
  | abort()ing pseudo client by server request. See https://wiki.yoctoproject.org/wiki/Pseudo_Abort for more details on this.
```

One potential culprit of this is in the openbmc_version.bbclass, which
is injected into every package.  In it is an invocation of `git` and
that invocation should be wrapped with a `export PSEUDO_DISABLE=1` so
that `pseudo` tracking is disabled during the `git` invocation.

I fixed a nearly identical problem in the upstream LF-OpenBMC tree
around a year ago[1] and have similar code in fb-source-mirror.bbclass
for similar reasons.

1. https://github.com/openbmc/openbmc/commit/bc442de08ff2e45ae01cb74397ccf010ef9797af

Test Plan: Built northdome machine locally.

Reviewed By: peterdelevoryas

Differential Revision: D30966638

fbshipit-source-id: 7019e4731a