openbmc-fb: split distro file into base
Summary:
The distro config for `openbmc-fb` contains a lot of content that is
typically found in a machine layer local.conf.sample, but are either
sane defaults for our machines or custom settings for the Facebook
build process.  Split parts of out it of `openbmc-fb` so that pieces
which are going to be used by both "LF-based" and "FB-based" machines
can be found in a common base distro definition.

Re-organize the distro config into smaller files based on their purpose.

Test Plan: Built an existing machine.

Reviewed By: amithash

Differential Revision: D31056506

fbshipit-source-id: 783911fdd6