i2c-tools: move to meta-backports and upgrade
Summary:
We previously had a duplicate version of the i2c-tools_3.1.2 recipe
in common with a custom patch for "block read in i2cget".  By having
a versioned recipe in the common layer, the package is effectively
pinned at an older version.

i2c-tools v4.3 has the support introduced by the custom patch and is
available in the current Yocto lf-master branch.  Move the recipe and
patches to meta-backports to stop shadowing the version found in Yocto.

   * rocko - copy recipe as-is from common to rocko.
   * zeus - copy recipe as-is from common to zeus.
   * dunfell - backport patches from 4.3 to 4.1.
   * hardknott - backport patches from 4.3 to 4.2.
   * master - use upstream 4.3 recipe as-is.

Test Plan: Built Northdome and Bletchley.

Reviewed By: amithash

Differential Revision: D31056501

fbshipit-source-id: 8ace80ac2a