oe-init-build-env: set COREROOT one dir higher for trees
Summary:
The Yocto trees with `lf-<branch>` have poky subtrees at the root of
the `yocto/<branch>`, but also have a `poky` subtree similar to our
old style Yocto imports.  Yocto itself doesn't care where the `poky`
subtree is but some of the LF recipes in `meta-phosphor` get confused
when there is an extra directory segment in `COREROOT`.

Pivot our COREROOT by removing the `/poky` portion of the environment
sourcing when pointing to an `lf-<branch>` version.  This also requires
reducing the corresponding path segment in all machine
`bblayers.conf.sample` files when using an `lf-<branch>` version.

Also, we have a useless `meta-yocto-bsp` layer being used in many
machines and that layer does not exist in the `lf-<branch>` trees (at
the root).  Rather than add an extra `poky` to these paths, remove the
layer tree-wide since it is not providing anything for us.

Test Plan: Build Northdome and Bletchley.

Reviewed By: amithash

Differential Revision: D31056509

fbshipit-source-id: 2e85ef2641