Make REST API testing script work on new platforms
Summary:
There's quite a discrepancy in how we call platforms in various
places, on example of ND:
- build name, like "fby2-nd"
- machine name (presented in /etc/issue), "northdome"

Moreover, fbnd is a subplatform of fby2, so there's additional fix
needed to determine the rest-api sources directory.

This is merely an attempt, haven't tested it fo all platforms yet.
Publishing to unblock xaionaro .

Test Plan: Tested on ND and Y2, more tests to come.

Reviewed By: deathowl

Differential Revision: D31055788

fbshipit-source-id: 99108e0d1f