Add failure logic for yamg fpga_ver
Summary:
The value of fpga_ver.sh should never return val1 as 0.
Therefore, we add a small logic to take care of that. See the attached
task for more details if needed T99694442

Test Plan:
eyeball since I didn't find a unit in that failed state
the output "2021-09-01_14:11:38.88371 PIM 1 : 0.0" should never produce
a pim with value 0.0, so we check val1 to see if it's 0. eyeball as
testing since we don't have a unit in that state right now.

Reviewed By: mikechoifb

Differential Revision: D31536118

fbshipit-source-id: f223e7e5e7