check FRU's presense before starting test
Summary: This is a followup diff to D32076574. when we try to power on a no exist FRU, power-util will exit with code 255 causing test exception. However in our devices in basset pool, we can't always guarantee to have full HW installed like we had in the production. So we'll have to first check HW availability before the test start, and only power-on devices which are marked as available.

Test Plan:
Sandcastle FBTP: https://www.internalfb.com/intern/sandcastle/job/4503600141224000
Sandcastle FBY3:
https://www.internalfb.com/intern/sandcastle/job/18014399021045635

Reviewed By: binhuang00

Differential Revision: D32108575

fbshipit-source-id: c158724f98