Revert "common: openbmc-utils, add ubifs_resize"
Summary:
This reverts commit ba057d7cc3259ff27ef475378cd4ce6940d04d58.

The discussed plan was to commit this in order to get it in a release
for Northdome and then revert the commit and revisit later.  We plan to
convert this to a stand-alone script which is not part of the boot path
on all BMCs.

Test Plan: Simple revert.

Reviewed By: zhdaniel12

Differential Revision: D31996213

fbshipit-source-id: 27a4732d40