rackmond2: Add UART device specialization
Summary:
Inheriting from `Device` we have `UARTDevice` which implements all the
UART specific methods. In particular it has the concept of baudrate,
enabling/disabling read. It also overrides `wait_write` method to actually
poll the device on write completion.

Inheriting from `UARTDevice` we have `RS485NativeDevice` which is the
abstraction for the native RS485 Device on the ASPEED Chip.

Test Plan: #doitlive

Reviewed By: garnermic

Differential Revision: D32051545

fbshipit-source-id: 691a047182