rackmond2: Add basic encode/decode message abstraction
Summary:
`msg.hpp` provides a native `Msg` structure which allows us to create packs/unpacks
very easily. It provides `<<` operator to allow one to encode fields and
`>>` operator to decode fields.
Currently only byte and half-byte is supported since most of the work revolves
around those. We can add more in the future if future messages need it.
It correctly uses big-endian for members larger than a byte.

`Command` is a pair of request and response which is consumed by `Modbus::command`.

Test Plan: #doitlive

Reviewed By: garnermic

Differential Revision: D32051542

fbshipit-source-id: 1aeec3370e