rackmond2: Specialize Msg to provide Modbus commands
Summary:
Add the defined modbus messages
Msg is further specialized in modbus_cmds.hpp to provide all
the defined modbus Request/Responses.
We are starting with adding ReadHoldingRegisters (Req/Resp)
since that is the basic necessary for monitoring and probing.

We can add the writes later when we bring in support for
baudrate negotiation and timestamp setting.

Test Plan: Build

Reviewed By: garnermic

Differential Revision: D32051541

fbshipit-source-id: 1f1c836e29