rackmond2: Add UNIX socket service layer (main)
Summary:
Add the abstraction to the encoding used between
the rackmond UNIX socket interface. This defines
the socket layer in the header (to be shared with
the client).

The idea is to move all communication from the previous
binary request and string response to a two-way JSON
request/response.

Since rest currently uses the legacy, this is maintaining
the legacy interface just for the raw-cmd message (TODO).
For everything else, it moves to the extendable JSON
interface.

Test Plan:
Build on rocko and dunfell.
Try out the service on devserver with the rackmon bits commented out.

Reviewed By: garnermic

Differential Revision: D32300423

fbshipit-source-id: d7ecc92f32