rackmond2: Fix CRC16 calculation
Summary:
This was a stupid mistake of using the fancy C++ for-loop
to iterate over the raw bytes. Unfortunately, this iterates
over the entire array (253 bytes) instead of the actual length.

Change to use the new iterators provided by the Msg class.

Test Plan: Test on hardware and ensure we can detect devices

Reviewed By: garnermic

Differential Revision: D32450372

fbshipit-source-id: e68d831918