rackmond: Make regmap table format configuration easy to read
Summary:
Instead of a map from bit name -> bit, use an array of
tuples (bit, name). This way, we do not have to worry about things
reordering the map based on name.
(Well, too bad JSON does not support maps with integers as keys).

Also allow "format" to be optional. If I dont care how its formatted,
I should not waste space in the configuration to specify it as hex.

Test Plan:
Build load on BMC and ensure nothing breaks.
Rebuild with debug prints and ensure the regmap is loaded correctly.

Reviewed By: garnermic

Differential Revision: D32477015

fbshipit-source-id: b0eafadff3