rest-api: remove native and support future distros
Summary:
The 'rest-api-native' version of this recipe is no longer used
for unit test, so simplify the recipe significantly by removing the
split native/target class variants.  Add a small distro-check so that
after Dunfell we start using the upstream `python-aiohttp` package
instead of our own versions.

Test Plan:
Built a handful of systems; CIT will cover the rest.  Should
have no affect to current systems as they are all on Dunfell or earlier.
Built Northdome with Yocto-Hardknott and tested in QEMU.

Reviewed By: doranand

Differential Revision: D33262717

fbshipit-source-id: 68a0bd22c7