rackmon2: Add support to return structured value data
Summary:
This adds full support to retrieve the value data
(Use of RegisterValue instead of the legacy raw-data
or strings).

This also does some changes to the function names to make
them all consistent. The fmt variants are on deprecation
path since this introduction.

Test Plan: run all unit tests.

Reviewed By: garnermic

Differential Revision: D33346392

fbshipit-source-id: d89f10dfc2