busybox: fix masking of lf config
Summary:
We have an DISTROOVERRIDE (openbmc-fb) to select busybox
customizations on those systems, but it was observed that those
overrides were still applying to openbmc-fb-lf machines.  This is
because the file name matches (busybox.cfg) and so Yocto will search in
our layers before the upstream layers.  Rename the config file to
`fb-busybox.cfg` to circumvent this behavior.

Test Plan: Built bletchley and northdome.  Checked busybox config.

Reviewed By: garnermic

Differential Revision: D33349568

fbshipit-source-id: ad33f7c7e5