fb-openbmc-image: handle upstream config
Summary:
Since LF-OpenBMC uses 'obmc-phosphor-image' and we use
'MACHINE-image', we have an indirection in 'fb-openbmc-image.inc'
to make this work.  The problem was that if there were any
'obmc-phosphor-image.bbappend' files in LF-OpenBMC's meta-facebook,
they would not apply.  In upstream's meta-facebook, I added an
'fb-phosphor-image.inc' and a very minimal
'obmc-phosphor-image.bbappend' so that we can include the same files
for machines built using this repo.

Test Plan: Built bletchley and checked 'bitbake -ne' output.

Reviewed By: garnermic

Differential Revision: D33346816

fbshipit-source-id: 9448d319df