circle-ci: qemu: Add firmware to secondary flash
Summary:
Up until now, we've only been providing the golden firmware image in the primary flash to QEMU. This means we were always booting in recovery mode, because the secondary flash was undefined, and wouldn't pass verified boot.

Since openbmc.qemu:7, we now support specifying the secondary flash firmware image as an additional -drive MTD argument. We still specify the same firmware file, but this allows verified boot to pass for unsigned images in QEMU, allowing us to run in non-recovery mode in QEMU. We still can't boot signed images in non-recovery mode because the TPM is not emulated at all, so all requests sent to the TPM are ignored and fail. But unsigned images don't require that the TPM verifies the image.

This is the same thing as the netcastle diff D33553185, but for CircleCI, because we don't have a mechanism to share code between netcastle and this repo. Ideally, we should find some way to share this code between the two repositories.

Test Plan: https://app.circleci.com/pipelines/github/facebook/openbmc/2552/workflows/b36cf00d-6324-4a75-adcb-d0be1cf193a6/jobs/56083

Reviewed By: williamspatrick

Differential Revision: D33587237

fbshipit-source-id: 784886e873