Modifying Fuji's peutil to take 2 based number, that is, 2-9 (from 1-8)
Summary:
[Fuji] Modifying Fuji's peutil to take 2 based number, that is, 2-9 (from 1-8)
We make the change to fix provisioning issue that is blocking the switch deployment.
Note that this change will modify the behavior of peutil of fuji only. The REST endpoint should remain the same as before (1-8), to prevent other FB infra relying on this from breaking.

Test Plan:
Build Pass.
peutil and related REST endpoint tested in MPK17 fuji as in : P476638906

Reviewed By: tao-ren

Differential Revision: D33802425

fbshipit-source-id: d2ddb4615e