rackmon2: Add capability to provide endian quirks of register
Summary:
Some registers (particularly ORv2 registers which map to BBUs)
are in little-endian format instead of the Big endian as specified
by the Modbus protocol. Thus, this diff adds the capability
for a register map to specify if something is little-endian so
the value conversion can reswap the bits around to form the
correct integer value.

Also add the correct unit tests.

Test Plan: All unit tests pass.

Reviewed By: garnermic

Differential Revision: D34094964

fbshipit-source-id: 5e680c997e