tests2: Add --denylist option to cit_runner.py (#180)
Summary:
`cit_runner.py` checks for the existence of the file `/usr/local/bin/tests2/dummy_qemu` to determine if we're in QEMU. That, in turn, determines whether we use `qemu_denylist.txt` to filter out tests.

This turns out to not work well in Netcastle, because even in QEMU Netcastle runs, we run `cit_runner.py` outside of QEMU to list tests. We need a better way to specify when to use the denylist.

I think the simplest thing is to just make it an optional argument: in CircleCI and QEMU Netcastle runs, we can specify the denylist path in the invocation, and everywhere else we can just ignore the denylist argument.

Pull Request resolved: https://github.com/facebook/openbmc/pull/180

Test Plan:
Watching CircleCI to make sure it doesn't break.
https://app.circleci.com/pipelines/github/facebook/openbmc/2716/workflows/db865f23-470d-43f6-80bd-6bbe697f5151
{F701903629}

Reviewed By: binhuang00

Differential Revision: D34281364

fbshipit-source-id: 521166ee27