circleci: Force qemux86 builds unconditionally
Summary:
I refactored the CircleCI jobs so that we use the existing job steps for qemux86, but I forgot that there's a short-circuiting step that will skip building. This seems to be triggering, and then there's no qemux86-image built, and the unit-test job fails (it tries to build the qemux86-image as root, and bitbake doesn't let you run as root).

{F702348874}

{F702348914}

{F702348939}

When I reran the job with ssh, I noticed the deploy directory is missing (no qemux86-image was built)

{F702348997}

Test Plan: Checking with David if this looks ok

Reviewed By: jerrt2003

Differential Revision: D34325189

fbshipit-source-id: b0c4d66d02