tests2: Add --tmpfs argument to ptest-runner (#182)
Summary:
In https://github.com/facebook/openbmc/issues/177, I changed ptest-runner to unpack the rootfs in the working directory, and then forced CircleCI to cd to /mnt/ramdisk, so that the rootfs would get unpacked in a tmpfs.

Instead, we can just specify the tmpfs directory as an optional argument for this situation, and keep the old behavior for everything else. This avoids leaving the rootfs sitting around in people's working directories too.

Pull Request resolved: https://github.com/facebook/openbmc/pull/182

Test Plan: Verifying CircleCI and Sandcastle ptest runs pass.

Reviewed By: amithash

Differential Revision: D34312308

fbshipit-source-id: 65908be58d