local-source: add sanity check for pseudo-abort issues
Summary:
Add a sanity check into the do_package step to determine if the package
has any potential pseudo-abort issues.  The following is required to
have a pseudo-abort issue:

   - Source is installed into WORKDIR directly (ie. S == WORKDIR).
   - INHIBIT_DEBUG_PACKAGE_SPLIT is not enabled.
   - The SRC_URI contains raw source files which are also included into
     the debug packages.

We don't want this in the general parsing pass because every recipe is
parsed even if we do not include it in an image.  There are a few
upstream recipes that have potential issues.

Test Plan:
* Built northdome and cmm without any warnings raised.
* Manually forced `bitbake usleep` and observed:

```
WARNING: usleep-1.0-r0 do_package: SRC_URI contains source files but S == WORKDIR; this has likely pseudo-abort issues: file://usleep.c
```

Reviewed By: amithash

Differential Revision: D34400212

fbshipit-source-id: 6fbd08f9bf