rackmon2: Gracefully handle unavailable interface
Summary:
Some wedge400s have /dev/ttyUSB0 disappear all of a sudden.
We need a backup plan in case that happens instead of having
rackmond crash at boot due to open fails.

The solution used is to try to wait till the device reappears.
If the devices goes away while it was active, we still have
proper exception handling.

TODO1: If an interface is not present at start, then probe might be really
slow to detect devices on that interface.

TODO2: If we truly need to handle a TTY from going down in the middle
and then reappearing, then we need to support a monitor thread
to detect and reinit.

Test Plan: #doitlive, all unit tests pass.

Reviewed By: tao-ren

Differential Revision: D34703629

fbshipit-source-id: 4a06e5d118