rest-api: fix resource path for aiohttp on Honister
Summary:
The result of casting an aiohttp Resource to a string changed with the
version of aiohttp included in Yocto Honister.  This previously caused
all of the Resource properties in the API to have trailing ">"
characters.  In order to ensure backwards compatibility with
pre-Honister, do a simple hack to strip off any trailing ">" characters.

Test Plan: Ran on Northdome with Honister and Dunfell.

Reviewed By: garnermic

Differential Revision: D35127439

fbshipit-source-id: c757480a485c18721b9bf391cca136c6fa96f082