dhcp: commonize ncsi patch
Summary:
Many machines have a special bbappend to patch dhcp with an NCSI
workaround.  Rather than have every machine require to bbappend the
package add a MACHINE_FEATURE which can trigger it and define a common
include.  This will allow us to automatically adapt to the right
behavior if/when we switch to another DHCP client.

Test Plan:
Build Northdome and Angelslanding and confirmed patch was
applied.

Reviewed By: GoldenBug

Differential Revision: D35151464

fbshipit-source-id: 4b6b8bf2919cff8a0e11d5558da59113174725e3