psu-update-delta: Work around timeout exception
Summary:
When we receive a timeout during a firmware upgrade, we can
potentially leave the device in a non-recoverable state. At this
point retrying send of the packet is worse since the PSU may
receive a duplicated packet. For example in the following flow
```
1. FOR segment in FIRMWARE_SEGMENTS:
2.     SET_WRITE_ADDRESS(segment.START_ADDRESS)
3.     FOR chunk in segment.data: # chunk is at most 8 bytes
4.         WRITE_DATA(chunk)
```
If we retry step-4 the PSU can mistake the repeated chunk for
the next chunk and can completely corrupt the flash block.

The alternative would be to repeat from step-2. But, in most
HW implementations, the flash block has to be erased before
written to. Hence the current workaround is.

a. Ignore timeout exception and allow the firmware upgrade flow
to continue. This assumes that the PSU received the packet
but we did not receive the response. If the PSU really did
not receive the packet, then we should fail in the final
verify step which would catch this step.

b. Put the above (Along with the erase and verify steps) in
a retry loop. That way, if we see timeouts, we ignore them.
and we head to the verify step. If we fail here, we retry
by issuing an erase.

Test Plan: Upgrade V2 and G2 PSU on wedge400.

Reviewed By: GoldenBug

Differential Revision: D35554339

fbshipit-source-id: 6a4e7f20d4ae40d1647f987fd375fc8336aa169c