fby35: Remove extraneous key pointing to temp data store
Summary:
As part of the ACD patch, a leftover #define was left in the code after review
which pointed to the temp data store.  Removing the key to keep in spec with
the requirement not to have any pointers to the kv temp or perm data stores
in the yv3.5 code.

Test Plan: Code built successfully without the #define

Reviewed By: williamspatrick

Differential Revision: D35825095

fbshipit-source-id: e97537152cf85ad5dbf799cbad49ac9023939a57