gpiolib-of: Respect High CS in spi-gpio (#245)
Summary:
This is needed to enable High CS to be respected in dts for spi-gpio
nodes.

On some dts, we have High CS, for example on fbdarwin with spi-gpio, where
the High CS is not respected to a logic error in the gpiolib-of.

This patch was sourced from here:
https://www.spinics.net/lists/linux-gpio/msg40354.html

X-link: https://github.com/facebookexternal/openbmc.arista/pull/245

Test Plan:
on fbdarwin, successfully read/write and root cause this issue using
logic analyser.

Reviewed By: joancaneus

Differential Revision: D35292205

fbshipit-source-id: 169209c774fc4c2e7ba6e0690b78efebf00ca01f