churchillmono: revamp proprietary recipes
Summary:
The churchillmono machine requires proprietary recipes which currently
point to private repositories that most developers and automation do
not have access to.  We are still working through the appropriate way
to store these recipes but in the meantime move the layer under
`proprietary/` so they are not shipped to the public github repo.
Also, disable the recipes by default for developers and automation but
allow specific developers to manually enable the layers in their
`build/conf/bblayers.conf` file.

Test Plan:
`bitbake churchillmono-image -ne` with BBLAYERS set both
ways to confirm `cisco-wb-kmod` was present/not-present as expected.

Reviewed By: binhuang00

Differential Revision: D36312644

fbshipit-source-id: d8afc3337f272f8975714f217924085d9f475cb5