bitbake: config changes for kirkstone
Summary:
Kirkstone has changed some potentially offensive variable
names and it is done in a way which is not backwards compatible.
Python2 hasn't been in the tree for a long time, so these blocks
are not necessary.  The busybox-syslog block is already handled
elsewhere.  And the image-prelink class hasn't done anything for
a while.

Test Plan: Build numerous platforms on various Yocto versions.

Reviewed By: doranand

Differential Revision: D36304553

fbshipit-source-id: 5205fe428383ec2043b8860d2a8b3c239998a27d