distro: force IPK packaging
Summary:
In some cases the weak-set in one of the Yocto configs was taking
precedence over ours and causing us to use RPM packaging.  This
was noticed as a compile failure whenever a machine did not use
the `meta-perl` layer.  Force-set to IPK since we want to use IPK
everywhere anyhow.

Test Plan: Successfully built halfdome with lf-master.

Reviewed By: amithash

Differential Revision: D36644358

fbshipit-source-id: 6412b1d3e16832edc904c532d917ca5178f59a86