emmc: kernel: reduce default modules
Summary:
We are previously enabling EXT4 and BTRFS as kernel modules on all
systems, but these file systems are only needed when eMMC is enabled.
Switch these from being specified in the defconfig to being triggered
via config snippets.  This allows non-eMMC systems to not include these
file systems, which frees up approximately 5MB uncompressed space in
the initramfs on those systems.

Test Plan:
Built a set of systems to confirm the config is added/not added as
expected.  `{churchillmono, grandcanyon, halfdome, yv35}`

Reviewed By: tao-ren

Differential Revision: D36668372

fbshipit-source-id: 98897798ee27213d8fbf5d0a46c7253764734a7d