zstd: refactor image variables
Summary:
A number of systems are missing the definitions in their
image in order to boot from a zstd-compressed rootfs.  This filesystem
type boots much quicker with almost no overhead compared with lzma.
Refactor how this is specified into an include-file to hopefully
reduce the number of machines which miss this feature.

Test Plan: Built northdome and confirmed zstd boot.

Reviewed By: tao-ren

Differential Revision: D36668369

fbshipit-source-id: 83074422aa0171ab05930d7be73053a669505642