create /run/lock if missing
Summary: On lf-dunfell with systemd /run/lock seems to be missing.

Test Plan:
Build wedge100 image.  Check that it's unaffected, i.e. that /run/lock is
present and that systemd-tmpfiles doesn't error out.

```
root@fboss8382003-oob:~# systemd-tmpfiles --create
[/etc/tmpfiles.d/lock.conf:1] Duplicate line for path "/run/lock", ignoring.
"/var/log" already exists and is not a directory.
"/var/tmp" already exists and is not a directory.
```

Reviewed By: deathowl

Differential Revision: D36777220

fbshipit-source-id: 451a5fe3883ebb4ae14e69231eb8a7da5a507e50